JAMES-2763 Startup check API should be more generic
