+++ server/container/guice/guice-common/src/test/java/org/apache/james/StartUpChecksPerformerTest.java
@@ -23,15 +23,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.apache.james.StartUpChecksPerformer.StartUpCheck.CheckResult;
-import org.apache.james.StartUpChecksPerformer.StartUpCheck.ResultType;
+import org.apache.james.lifecycle.api.StartUpCheck;
+import org.apache.james.lifecycle.api.StartUpCheck.CheckResult;
+import org.apache.james.lifecycle.api.StartUpCheck.ResultType;
 import org.junit.jupiter.api.Test;
 
 class StartUpChecksPerformerTest {
 
     private static final CheckResult GOOD_CHECK_1 = CheckResult.builder()
         .checkName("good 1")
-        .resultType(ResultType.GOOD)
+        .resultType(StartUpCheck.ResultType.GOOD)
         .build();
     private static final CheckResult GOOD_CHECK_2 = CheckResult.builder()
         .checkName("good 2")
@@ -46,8 +47,8 @@ class StartUpChecksPerformerTest {
         .resultType(ResultType.BAD)
         .build();
 
-    private static StartUpChecksPerformer.StartUpCheck fromResult(CheckResult checkResult) {
-        return new StartUpChecksPerformer.StartUpCheck() {
+    private static StartUpCheck fromResult(CheckResult checkResult) {
+        return new StartUpCheck() {
 
             @Override
             public CheckResult check() {
@@ -110,7 +111,7 @@ class StartUpChecksPerformerTest {
         String checkName = "throwing check name";
         StartUpChecksPerformer checksPerformer = StartUpChecksPerformer.from(
 
-            new StartUpChecksPerformer.StartUpCheck() {
+            new StartUpCheck() {
                 @Override
                 public CheckResult check() {
                     throw new RuntimeException("unexpected");
